Misc changes
* tutorials/sndkit/Makefile - make semi-decent Makefile to compile most subdirs.

* Remove silly casts from morse programs, and make scoring in morse2/3 a bit more accurate.

* Add compile time option to charlist.h to get rid of international letters (not on most keyboards).

* Remove warnings on softsynth, and add clean target to userdev Makefile.

* oss_sblive.man - sblive_digital_din description implied wrong default.
                 - Mention that audigys have a different default.
                 - Change "SBLIVE DEVICE USAGE" section name to "SBLIVE HARDWAR                    MIXING" name
                   (as now vmix is default anyway, and I'm not sure the
                    quality is lower with it), and update test.
                 - Enforce 80-col limit.

* ossmix.c - style - set oss_mixer_val at start of loop, and reset value member.
		Does not really change anything but code looks cleaner.

* os_linux.h/os_freebsd.h - use modern __i386__ instead of i386, same for __x86_64__.

+++ tutorials/sndkit/morse/morse.c
@@ -241,7 +241,7 @@ main (int argc, char *argv[])
     }
 
   parm = AFMT_S16_LE;
-  if (ioctl (audiofd, SNDCTL_DSP_SETFMT, (int) &parm) == -1)
+  if (ioctl (audiofd, SNDCTL_DSP_SETFMT, &parm) == -1)
     {
       perror ("SETFMT");
       close (audiofd);
@@ -258,7 +258,7 @@ main (int argc, char *argv[])
     }
 
   parm = SRATE;
-  if (ioctl (audiofd, SNDCTL_DSP_SPEED, (int) &parm) == -1)
+  if (ioctl (audiofd, SNDCTL_DSP_SPEED, &parm) == -1)
     {
       perror ("SPEED");
       close (audiofd);
